+++ src/gameserver/core/framework/Bundle.ts
@@ -1,0 +1,6 @@
+import { IBundle } from './BundleLoader';
+import { Module } from './ModuleLoaderStrategies';
+
+export interface Bundle extends IBundle {
+    modules: Module[],
+}