+++ src/gameserver/core/framework/ModuleLoaderStrategies/Module.ts
@@ -1,0 +1,7 @@
+import { IServiceModule } from './ServiceModuleLoaderStrategy';
+import { ISystemModule } from './SystemModuleLoaderStrategy';
+
+export type Module =
+    IServiceModule<any>
+    | ISystemModule<any>
+    ;