+++ src/gameserver/core/framework/ModuleLoaderStrategies/ServiceModuleLoaderStrategy/IServiceModule.ts
@@ -6,7 +6,7 @@ import {
 
 export interface IServiceModule<T> extends IModule<'service'> {
     type: 'service',
-    name: String,
+    name: string,
     factory: (locator: ISystemLocator) => T,
     identifiers: [TypedIdentifier<Partial<T>>, ...TypedIdentifier<Partial<T>>[]],
 }