+++ src/gameserver/core/framework/ModuleLoaderStrategies/index.ts
@@ -1,4 +1,5 @@
 export * as moduleTypes from './moduleTypes';
+export { Module } from './Module';
 
 export { ServiceModuleLoaderStrategy } from './ServiceModuleLoaderStrategy';
 export { SystemModuleLoaderStrategy } from './SystemModuleLoaderStrategy';