+++ src/gameserver/core/framework/ServiceManager/IServiceManager.ts
@@ -1,0 +1,6 @@
+import { TypedIdentifier } from '../DependencyLocator';
+
+export interface IServiceManager {
+    initializeAllServices(): void;
+    notifyServiceRegistered(identifier: TypedIdentifier<any>): void;
+}