+++ src/gameserver/core/framework/ServiceManager/index.ts
@@ -1,0 +1,2 @@
+export { IServiceManager } from './IServiceManager';
+export { ServiceManager } from './impl/ServiceManager';