+++ src/gameserver/core/framework/SystemManager/ISystemManager.ts
@@ -1,0 +1,6 @@
+import { TypedIdentifier } from '../DependencyLocator';
+
+export interface ISystemManager {
+    initializeAllSystems(): void;
+    notifySystemRegistered(identifier: TypedIdentifier<any>): void;
+}