+++ src/gameserver/core/framework/SystemManager/index.ts
@@ -1,0 +1,2 @@
+export { ISystemManager } from './ISystemManager';
+export { SystemManager } from './impl/SystemManager';