+++ src/gameserver/core/framework/index.ts
@@ -1,1 +1,3 @@
 export { moduleTypes } from './ModuleLoaderStrategies';
+export { Bundle } from './Bundle';
+export { Framework } from './Framework';