+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Game Server `BundleLoader`
 - Game Server `SystemModuleLoaderStrategy`
 - Game Server `ServiceModuleLoaderStrategy`
+- Game Server `SystemManager`
+- Game Server `ServiceManager`
+- Game Server `Framework`
 
 ### Changed
 ### Deprecated